+++ test/node/Replication.test.js
@@ -144,10 +144,10 @@ describe('Replication.test.js', () => {
 
     describe('events', () => {
         describe('positive', () => {
-            it('collection: should get an event when a docs syncs', async() => {
+            it('collection: should get an event when a doc syncs', async() => {
                 const serverURL = await SpawnServer.spawn();
-                const c = await humansCollection.create(0, 'colSource ' + randomToken(5));
-                const c2 = await humansCollection.create(0, 'colSync ' + randomToken(5));
+                const c = await humansCollection.create(0, 'colsource' + randomToken(5));
+                const c2 = await humansCollection.create(0, 'colsync' + randomToken(5));
                 c.sync(serverURL, {
                     live: true
                 });
@@ -175,8 +175,8 @@ describe('Replication.test.js', () => {
 
             it('query: should re-find when a docs syncs', async() => {
                 const serverURL = await SpawnServer.spawn();
-                const c = await humansCollection.create(0, 'colSource ' + randomToken(5));
-                const c2 = await humansCollection.create(0, 'colSync ' + randomToken(5));
+                const c = await humansCollection.create(0, 'colsource' + randomToken(5));
+                const c2 = await humansCollection.create(0, 'colsync' + randomToken(5));
                 c.sync(serverURL, {
                     live: true
                 });
@@ -205,8 +205,8 @@ describe('Replication.test.js', () => {
             });
             it('document: should change field when doc saves', async() => {
                 const serverURL = await SpawnServer.spawn();
-                const c = await humansCollection.create(0, 'colSource ' + randomToken(5));
-                const c2 = await humansCollection.create(0, 'colSync ' + randomToken(5));
+                const c = await humansCollection.create(0, 'colsource' + randomToken(5));
+                const c2 = await humansCollection.create(0, 'colsync' + randomToken(5));
                 c.sync(serverURL, {
                     live: true
                 });